backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
langsmith>=0.0.30
PyPDF2>=3.0.0
python-docx>=0.8.11
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0